Rework console.js into LiveView Hook for UI 2.0, organize app.js and existing hooks
