+++ assets/ui-rework/app.js
@@ -2,380 +2,43 @@ import "phoenix_html"
 import { Socket } from "phoenix"
 import { LiveSocket } from "phoenix_live_view"
 import topbar from "topbar"
-import L from "leaflet/dist/leaflet.js"
 import { createLiveToastHook } from "live_toast"
-import Chart from "chart.js/auto"
 import "chartjs-adapter-date-fns"
-import "leaflet.markercluster/dist/leaflet.markercluster.js"
-import mapboxgl from "mapbox-gl"
-
-import hljs from "highlight.js/lib/core"
-import bash from "highlight.js/lib/languages/bash"
-import elixir from "highlight.js/lib/languages/elixir"
-import plaintext from "highlight.js/lib/languages/plaintext"
-import shell from "highlight.js/lib/languages/shell"
-hljs.registerLanguage("bash", bash)
-hljs.registerLanguage("elixir", elixir)
-hljs.registerLanguage("plaintext", plaintext)
-hljs.registerLanguage("shell", shell)
 
 import TimeAgo from "javascript-time-ago"
 import en from "javascript-time-ago/locale/en"
 
+import Console from "./hooks/console.js"
+import SharedSecretClipboardClick from "./hooks/sharedSecretClipboardClick.js"
+import Chart from "./hooks/chart.js"
+import HighlightCode from "./hooks/highlightCode.js"
+import UpdatingTimeAgo from "./hooks/updatingTimeAgo.js"
+import LocalTime from "./hooks/localTime.js"
+import SimpleDate from "./hooks/simpleDate.js"
+import WorldMap from "./hooks/worldMap.js"
+import DeviceLocationMap from "./hooks/deviceLocationMap.js"
+
 import dates from "../js/dates"
 
 TimeAgo.addDefaultLocale(en)
-
-let Hooks = {
-  LiveToast: createLiveToastHook()
-}
-
-Hooks.DeviceLocationMap = {
-  mounted() {
-    let accessToken = this.el.dataset.accessToken
-    let style = this.el.dataset.style
-    let centerLng = this.el.dataset.centerLng
-    let centerLat = this.el.dataset.centerLat
-    let zoom = this.el.dataset.zoom
-    let source = this.el.dataset.source
-
-    const ctx = this.el
-
-    mapboxgl.accessToken = accessToken
-
-    const map = new mapboxgl.Map({
-      container: ctx,
-      style: style,
-      center: [centerLng, centerLat],
-      zoom: zoom
-    })
-
-    map.addControl(new mapboxgl.NavigationControl())
-
-    if ((source || "").toLowerCase() == "gps") {
-      new mapboxgl.Marker({ color: "rgb(99 102 241)" })
-        .setLngLat([centerLng, centerLat])
-        .addTo(map)
-    }
-  }
-}
-
-Hooks.SharedSecretClipboardClick = {
-  mounted() {
-    const parent = this.el
-    this.el.addEventListener("click", () => {
-      const secret = document.getElementById("shared-secret-" + parent.value)
-        .value
-      if (typeof ClipboardItem && navigator.clipboard.write) {
-        // NOTE: Safari locks down the clipboard API to only work when triggered
-        //   by a direct user interaction. You can't use it async in a promise.
-        //   But! You can wrap the promise in a ClipboardItem, and give that to
-        //   the clipboard API.
-        //   Found this on https://developer.apple.com/forums/thread/691873
-
-        const type = "text/plain"
-        const blob = new Blob([secret], { type })
-        const data = [new window.ClipboardItem({ [type]: blob })]
-        navigator.clipboard.write(data)
-
-        confirm("Secret copied to your clipboard")
-      } else {
-        // NOTE: Firefox has support for ClipboardItem and navigator.clipboard.write,
-        //   but those are behind `dom.events.asyncClipboard.clipboardItem` preference.
-        //   Good news is that other than Safari, Firefox does not care about
-        //   Clipboard API being used async in a Promise.
-        navigator.clipboard.writeText(secret)
-        confirm("Secret copied to your clipboard")
-      }
-    })
-  }
-}
-
-Hooks.Chart = {
-  dataset() {
-    return JSON.parse(this.el.dataset.metrics)
-  },
-  unit() {
-    return JSON.parse(this.el.dataset.unit)
-  },
-  mounted() {
-    let metrics = JSON.parse(this.el.dataset.metrics)
-    let type = JSON.parse(this.el.dataset.type)
-    let max = JSON.parse(this.el.dataset.max)
-    let min = JSON.parse(this.el.dataset.min)
-    let title = JSON.parse(this.el.dataset.title)
-
-    const ctx = this.el
-    var data = []
-    for (let i = 0; i < metrics.length; i++) {
-      data.push(metrics[i])
-    }
-
-    const areaChartDataset = {
-      type: "line",
-      data: {
-        datasets: [
-          {
-            backgroundColor: "rgba(99, 102, 241)",
-            fill: {
-              target: "start",
-              above: "rgba(99, 102, 241, 0.29)",
-              below: "rgba(99, 102, 241, 0.29)"
-            },
-            radius: 2,
-            data: this.dataset()
-          }
-        ]
-      },
-      options: {
-        plugins: {
-          title: {
-            display: true,
-            align: "start",
-            text: title,
-            font: {
-              size: 16,
-              weight: "normal"
-            },
-            color: "rgba(212, 212, 216)",
-            padding: {
-              bottom: 14
-            }
-          },
-          legend: {
-            display: false
-          }
-        },
-        scales: {
-          x: {
-            grid: {
-              color: "rgba(63, 63, 70)"
-            },
-            type: "time",
-            time: {
-              unit: this.unit(),
-              displayFormats: {
-                millisecond: "HH:mm:ss.SSS",
-                second: "HH:mm:ss",
-                minute: "HH:mm",
-                hour: "HH:mm"
-              }
-            },
-            ticks: {
-              display: true,
-              autoSkip: false
-            }
-          },
-          y: {
-            offset: true,
-            grid: {
-              color: "rgba(63 63 70)"
-            },
-            type: "linear",
-            min: 0
-            // max: max
-          }
-        },
-        responsive: true,
-        maintainAspectRatio: false
-      }
-    }
-
-    const chart = new Chart(ctx, areaChartDataset)
-    this.el.chart = chart
-
-    this.handleEvent("update-charts", function(payload) {
-      if (payload.type == type) {
-        chart.data.datasets[0].data = payload.data
-        chart.update()
-      }
-    })
-
-    this.handleEvent("update-time-unit", function(payload) {
-      chart.options.scales.x.time.unit = payload.unit
-      chart.update()
-    })
-  },
-  updated() {}
-}
-
-Hooks.HighlightCode = {
-  mounted() {
-    this.updated()
-  },
-  updated() {
-    hljs.highlightElement(this.el)
-  }
-}
-
-Hooks.UpdatingTimeAgo = {
-  updateTimer: null,
-  mounted() {
-    this.updateTimer = null
-    this.updated()
-  },
-  updated(element) {
-    let hook = arguments.length > 0 ? element : this
-
-    if (hook.updateTimer) {
-      clearTimeout(hook.updateTimer)
-    }
-
-    const timeAgo = new TimeAgo("en-US")
-
-    let dtString = hook.el.dateTime
-    let dt = new Date(dtString)
-
-    const formattedDate = timeAgo.format(dt, "round-minute")
-
-    hook.el.textContent = formattedDate
-
-    // https://www.npmjs.com/package/javascript-time-ago#update-interval
-    // set update interval to 10sec
-    let interval = 10000
-
-    hook.updateTimer = setTimeout(hook.updated, interval, hook)
-  }
-}
-
-Hooks.LocalTime = {
-  mounted() {
-    this.updated()
-  },
-  updated() {
-    let dt = new Date(this.el.textContent.trim())
-
-    dt.setSeconds(null)
-
-    let formatted = new Intl.DateTimeFormat("en-GB", {
-      day: "numeric",
-      month: "short",
-      year: "numeric",
-      hour: "numeric",
-      minute: "numeric",
-      hourCycle: "h12",
-      timeZone: Intl.DateTimeFormat().resolvedOptions().timeZone
-    }).format(dt)
-
-    this.el.textContent = formatted
-  }
-}
-
-Hooks.SimpleDate = {
-  mounted() {
-    this.updated()
-  },
-  updated() {
-    this.el.textContent = dates.formatDate(this.el.textContent)
-  }
-}
-
-Hooks.WorldMap = {
-  mounted() {
-    let self = this
-    let mapId = this.el.id
-    this.markers = []
-
-    var mapOptionsZoom = {
-      attributionControl: false,
-      zoomControl: true,
-      scrollWheelZoom: true,
-      boxZoom: false,
-      doubleClickZoom: false,
-      dragging: true,
-      keyboard: false,
-      maxZoom: 18,
-      minZoom: 1.4,
-      renderer: L.canvas()
-    }
-
-    var mapStyle = {
-      stroke: true,
-      color: "#2A2D30",
-      fillColor: "#b7bec5",
-      weight: 0.5,
-      opacity: 1,
-      fillOpacity: 0.5
-    }
-
-    // initialize the map
-    this.map = L.map(mapId, mapOptionsZoom).setView([0, 0], 1)
-    this.map.setMaxBounds(this.map.getBounds())
-    this.handleEvent("markers", ({ markers }) => {
-      self.markers = markers
-      self.updated()
-    })
-
-    // load GeoJSON from an external file
-    fetch("/geo/world.geojson")
-      .then(res => res.json())
-      .then(data => {
-        L.geoJson(data, { style: mapStyle }).addTo(this.map)
-        this.pushEvent("map_ready", {})
-      })
-  },
-  updated() {
-    let mode = this.el.dataset.mode
-
-    var myRenderer = L.canvas({ padding: 0.5 })
-    var clusterLayer = L.markerClusterGroup({
-      chunkedLoading: true,
-      chunkProgress: this.updateProgressBar
-    })
-
-    var defaultOptions = {
-      radius: 6,
-      weight: 1,
-      opacity: 0,
-      fillOpacity: 1,
-      renderer: myRenderer,
-      fillColor: "#4dd54f"
-    }
-
-    var offlineOptions = Object.assign(defaultOptions, {
-      fillColor: "rgba(196,49,49,1)"
-    })
-    var outdatedOptions = Object.assign(defaultOptions, {
-      fillColor: "rgba(99,99,99,1)"
-    })
-
-    var devices = this.markers.reduce(function(acc, marker) {
-      let location = marker["l"]
-      var latLng = [location["at"], location["ng"]]
-
-      // if no location or we don't care about mode, move on
-      if (!location["ng"] && !location["at"]) return acc
-      if (!["connected", "updated"].includes(mode)) return acc
-
-      if (mode == "connected") {
-        if (marker["s"] == "connected") {
-          acc.push(L.circleMarker(latLng, defaultOptions))
-        } else {
-          acc.push(L.circleMarker(latLng, offlineOptions))
-        }
-      }
-
-      if (mode == "updated") {
-        if (marker["lf"]) {
-          acc.push(L.circleMarker(latLng, defaultOptions))
-        } else {
-          acc.push(L.circleMarker(latLng, outdatedOptions))
-        }
-      }
-      return acc
-    }, [])
-
-    clusterLayer.addLayers(devices)
-    this.map.addLayer(clusterLayer)
-  }
-}
 
 let csrfToken = document
   .querySelector("meta[name='csrf-token']")
   .getAttribute("content")
 let liveSocket = new LiveSocket("/live", Socket, {
   params: { _csrf_token: csrfToken },
-  hooks: Hooks
+  hooks: {
+    Console,
+    SharedSecretClipboardClick,
+    Chart,
+    HighlightCode,
+    UpdatingTimeAgo,
+    LocalTime,
+    SimpleDate,
+    WorldMap,
+    DeviceLocationMap,
+    LiveToast: createLiveToastHook()
+  }
 })
 
 // Show progress bar on live navigation and form submits
