+++ assets/ui-rework/hooks/localTime.js
@@ -1,0 +1,22 @@
+export default {
+  mounted() {
+    this.updated()
+  },
+  updated() {
+    let dt = new Date(this.el.textContent.trim())
+
+    dt.setSeconds(null)
+
+    let formatted = new Intl.DateTimeFormat("en-GB", {
+      day: "numeric",
+      month: "short",
+      year: "numeric",
+      hour: "numeric",
+      minute: "numeric",
+      hourCycle: "h12",
+      timeZone: Intl.DateTimeFormat().resolvedOptions().timeZone
+    }).format(dt)
+
+    this.el.textContent = formatted
+  }
+}
