+++ assets/ui-rework/hooks/simpleDate.js
@@ -1,0 +1,8 @@
+export default {
+  mounted() {
+    this.updated()
+  },
+  updated() {
+    this.el.textContent = dates.formatDate(this.el.textContent)
+  }
+}
